% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticanalysis.r
\name{detect_dependencies}
\alias{detect_dependencies}
\title{Lists all packages that get imported by an R script.
Detects packages that are imported using library(package)
as well as with package::foo}
\usage{
detect_dependencies(file, recursive = TRUE)
}
\arguments{
\item{recursive}{Detect R scripts that are called using
source(script.R) and includes them in the analysis}
}
\value{
Character vector of imported package names
}
\description{
Lists all packages that get imported by an R script.
Detects packages that are imported using library(package)
as well as with package::foo
}
