% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{dotrepl}
\alias{dotrepl}
\title{dots in string are replaced with \\\\. to make them regex save.}
\usage{
dotrepl(s)
}
\arguments{
\item{s}{String}
}
\value{
String
}
\description{
dots in string are replaced with \\\\. to make them regex save.
}
