% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{get_dist_files}
\alias{get_dist_files}
\title{Get list of files to include in distribution package}
\usage{
get_dist_files(
  path = getwd(),
  ignore_start = c(),
  ignore_end = c(),
  dotreplace = T
)
}
\arguments{
\item{path}{Path}

\item{ignore_start}{Ignore paths that start with (e.g. c("plot") would exclude plot folder)}

\item{ignore_end}{Ignore paths that end with (e.g. c(".pdf",".docx") would ignore pdfs and word documents)}

\item{dotreplace}{dots in ignore list are replaced with \\\\. to make them regex save.}
}
\value{
list of files to include in distribution package
}
\description{
Get list of files to include in distribution package
}
