% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchinfo.R
\name{matchinfo}
\alias{matchinfo}
\title{Compare two ID-vectors. Shows duplicates and elements that don't match.}
\usage{
matchinfo(a, b, re = F)
}
\arguments{
\item{a}{Vector 1}

\item{b}{Vector 2}

\item{re}{(Optional) return results as a list? For example to use in unit testing.}
}
\value{
if re is set to true a list is returned.
}
\description{
Compare two ID-vectors. Shows duplicates and elements that don't match.
}
\examples{
matchinfo(c("a","b","c"), c("a","b","d","d"))
}
