% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticanalysis.r
\name{report_dependencies}
\alias{report_dependencies}
\title{Lists all packages that get imported by an R script.
Detects packages that are imported using library(package) 
as well as with package::foo}
\usage{
report_dependencies(file, recursive = TRUE)
}
\arguments{
\item{recursive}{Detect R scripts that are called using
source(script.R) and includes them in the analysis}
}
\value{
data.frame with attached version information for each package
}
\description{
Lists all packages that get imported by an R script.
Detects packages that are imported using library(package) 
as well as with package::foo
}
