% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{touch_dir}
\alias{touch_dir}
\title{Create folder if it does not exist.}
\usage{
touch_dir(foldername, dir = getwd())
}
\arguments{
\item{foldername}{Name of the folder}

\item{dir}{Directory to create the folder in}
}
\description{
Create folder if it does not exist.
}
