% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{touch_file}
\alias{touch_file}
\title{Create folder if it does not exist.}
\usage{
touch_file(filename, dir = getwd(), content = "")
}
\arguments{
\item{filename}{Name of the file}

\item{dir}{Directory to create the file in}

\item{content}{File content}
}
\description{
Create folder if it does not exist.
}
