% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{touch_script}
\alias{touch_script}
\title{Create script file if it does not exist.}
\usage{
touch_script(filename, dir = getwd(), content = "")
}
\arguments{
\item{filename}{Name of the script file}

\item{dir}{Directory to create the script in}

\item{content}{Script content}
}
\description{
Create script file if it does not exist.
}
