% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initproject.r
\name{get_dist_files}
\alias{get_dist_files}
\title{Get list of files to include in distribution package.}
\usage{
get_dist_files(
  path = getwd(),
  ignore_start = c(),
  ignore_end = c(),
  ignore_full = c(),
  dotreplace = T
)
}
\arguments{
\item{path}{Path}

\item{ignore_start}{Ignore paths that start with this}

\item{ignore_end}{Ignore paths that are exactly this (e.g. c("plot") would exclude plot folder)}

\item{dotreplace}{dots in ignore list are replaced with \\\\. to make them regex save.}
}
\value{
list of files to include in distribution package
}
\description{
Note: Due to limitations of the "zip" package, all files in subfolders will be included and can not be filtered/ignored.
}
